Add Archlinux install details to README
This is a fun little program. Congrats! I went ahead and bundled it for the Arch User Repository, and thought I'd go ahead and add the instructions to your readme, so you wouldn't have to, but it won't hurt my feelings if you didn't want to include this; Arch users can usually figure things out for themselves. 
The dependencies are a little confusing to me, but I'll open an issue about that.